requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
coverage==7.6.7
cryptography==43.0.3
cymem==2.0.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hypothesis==6.119.3
identify==2.6.1
idna==3.10
iniconfig==2.0.0
instructor==1.6.3
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jwt==1.3.1
langcodes==3.4.1
language_data==1.2.0
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
memory-profiler==0.61.0
multidict==6.1.0
murmurhash==1.0.10
mypy==1.13.0
mypy-extensions==1.0.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.0.2
openai==1.55.0
optional-django==0.1.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
preshed==3.0.9
propcache==0.2.0
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.41.2
swarm @ git+ssh://git@github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tomli==2.0.2
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yarl==1.17.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-jose==3.3.0
python-multipart==0.0.6
aiosmtplib==2.0.1
